+++ web-registry/src/components/molecules/index.ts
@@ -6,6 +6,8 @@ export * from './OverviewCards';
 export * from './HeroTitle';
 export * from './HeroAction';
 export * from './ImageItemsSection';
+export * from './LineItem';
+export * from './LineItemLabelAbove';
 export * from './TwoImageSection';
 export * from './PracticesOutcomesSection';
 export * from './TimelineSection';
